src/01.raw/igdb/requirements.txt
requests
boto3
python-dotenv
tqdm